+++ distributed/comm/tests/test_comms.py
@@ -14,7 +14,7 @@ from distributed.core import pingpong
 from distributed.metrics import time
 from distributed.utils import get_ip, get_ipv6
 from distributed.utils_test import (slow, loop, gen_test, gen_cluster,
-                                    requires_ipv6, has_ipv6, ssl_config)
+                                    requires_ipv6, has_ipv6, ssl_config, ssl_config_no_verify)
 
 from distributed.protocol import (loads, dumps,
                                   to_serialize, Serialized, serialize, deserialize)
@@ -780,7 +780,11 @@ def test_zmq_deserialize():
 def test_inproc_deserialize():
     yield check_deserialize('inproc://')
 
-@gen_test()
-def test_tls_deserializae(ssl_config):
 
-    yield check_deserialize('tls://')
+def test_tls_deserializae(loop, ssl_config_no_verify):
+
+    @gen_test()
+    def inner():
+        check_deserialize('tls://')
+
+    loop.run_sync(inner)
