+++ distributed/tests/test_ssl.py
@@ -1,24 +1,9 @@
 from __future__ import print_function, division, absolute_import
 
-from distributed.utils_test import gen_cluster, inc, loop, ssl_config
-import distributed.comm.utils as comm_utils
-
-import ssl
+from distributed.utils_test import gen_cluster, inc, loop, ssl_config_no_verify
 
 
-old_ssl_ctx_fx = comm_utils.create_ssl_context
-
-
-def create_ssl_context_mock():
-    ctx = old_ssl_ctx_fx()
-    ctx.check_hostname = False
-    # ssl_ctx.verify_flags =
-    ctx.verify_mode = ssl.CERT_NONE
-    return ctx
-
-
-def test_ssl(ssl_config, loop, monkeypatch):
-    monkeypatch.setattr(comm_utils, "create_ssl_context", create_ssl_context_mock)
+def test_ssl(ssl_config_no_verify, loop):
 
     @gen_cluster(
         client=True)
